+++ randomizer-data/src/main/java/de/alpharogroup/random/SecureRandomBuilder.java
@@ -78,10 +78,9 @@ public final class SecureRandomBuilder
 	/**
 	 * Gets an instance of {@link SecureRandomBuilder} from the given algorithm and provider
 	 *
-	 * @param algorithm
-	 *            the algorithm
-	 * @param provider
-	 *            the provider
+	 * @param algorithm the algorithm
+	 * @param provider  the provider
+	 * @param seed      the seed
 	 * @return the new {@link SecureRandomBuilder} object
 	 */
 	public static SecureRandomBuilder getInstance(final String algorithm, final String provider,